Turn all double into float; align everything b/w CPP/C (except a08-1 => acne ?)

+++ rtiow/C/a06-2.c
@@ -33,7 +33,7 @@ typedef struct hittable_s {
 bool list_hit(hittable *p, const ray *r, float t_min, float t_max, hit_record *rec) {
 	hit_record temp_rec;
 	bool hit_anything = false;
-	double closest_so_far = t_max;
+	float closest_so_far = t_max;
 	while (1) {
 		p++;
 		if (!p->hit) break;
